chore: module refactorisation
