+++ src/artefactual/calibration/rates_answers.py
@@ -9,7 +9,7 @@ import pandas as pd
 from pydantic import BaseModel, ValidationError
 from vllm import SamplingParams
 
-from artefactual.utils.models import init_llm
+from artefactual.calibration.helpers.models import init_llm
 
 logger = logging.getLogger(__name__)
